Clase de Conexiones agregada.
